refactor: move animated tooltip to file and add ondone
